extract structs from message.rs to its own files
