+++ rust/sglang-server/src/api_server/frame.rs
@@ -4,7 +4,7 @@
 //! abort frames). No HTTP here — the sibling `native_api` module owns the handlers
 //! and streams; it calls these per frame.
 
-use crate::message::{ChunkEvent, ChunkExtras};
+use crate::message::egress::{ChunkEvent, ChunkExtras};
 
 /// The text slot of a `[logprob, token_id, text]` tuple: the decoded token when
 /// `return_text_in_logprobs` supplied a text buffer, else `null`.
