+++ rust/sglang-server/src/api_server/native_api.rs
@@ -31,9 +31,10 @@ use super::frame::{
 };
 use super::guard::AbortGuard;
 use super::submit::submit;
-use crate::message::{
-    ChunkEvent, EgressItem, GenerateBody, GenerateRequest, RequestKind, SamplingParams, ids::Rid,
-};
+use crate::message::egress::{EgressItem, ChunkEvent};
+use crate::message::ids::Rid;
+use crate::message::request::{GenerateBody, GenerateRequest, RequestKind};
+use crate::message::sampling::SamplingParams;
 use crate::utils::{
     environ,
     response::{error_response, error_value},
@@ -540,7 +541,7 @@ fn terminal_stream_frame_string(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::message::ChunkEvent;
+    use crate::message::egress::ChunkEvent;
     use crate::tokenizer_manager::Senders;
     use crate::utils::error::Error;
     use futures::StreamExt;
