+++ rust/sglang-server/src/api_server/openai/tools.rs
@@ -36,7 +36,8 @@ use dynamo_protocols::types::{
     Role,
 };
 
-use crate::message::{ChunkEvent, SamplingParams};
+use crate::message::egress::ChunkEvent;
+use crate::message::sampling::SamplingParams;
 
 /// Canonicalize a tool-call parser name onto the dynamo-parsers registry keys.
 ///
@@ -263,7 +264,8 @@ mod tests {
         apply_tool_constraint, chat_delta, chat_finish_reason, dynamo_parser_name,
         dynamo_tool_choice, parse_chat_tool_calls,
     };
-    use crate::message::{ChunkEvent, SamplingParams};
+    use crate::message::egress::ChunkEvent;
+    use crate::message::sampling::SamplingParams;
     use dynamo_parsers::tool_calling::jail::{Annotated, apply_tool_calling_jail};
     use dynamo_parsers::{ToolChoice as DynamoToolChoice, ToolDefinition};
     use dynamo_protocols::types::CreateChatCompletionStreamResponse as StreamResponse;
