+++ rust/sglang-server/src/multi_modality/payload.rs
@@ -9,7 +9,7 @@ use bytes::Bytes;
 use rmpv::Value;
 use sglang_mm::driver::{ImageSource, MmInput};
 
-use crate::message::MmWorkItem;
+use crate::message::request::MmWorkItem;
 
 /// True for sources the API layer must resolve before MM dispatch: I/O — network
 /// *or* disk, since a network mount can hang past any HTTP timeout — never runs
