+++ rust/sglang-server/src/multi_modality/worker.rs
@@ -4,8 +4,8 @@
 use std::sync::Arc;
 
 use super::sidecar::{FeatureStore, MmSidecarEntry, Sidecar, park_features_in_shm};
-use crate::message::MmRequest;
 use crate::message::ids::Rid;
+use crate::message::request::MmRequest;
 use crate::tokenizer_manager::TmEvent;
 use crate::tokenizer_manager::tokenizer::TextTokenizer;
 use crate::utils::runtime::Runnable;
@@ -80,7 +80,7 @@ impl Context {
 fn process(
     ctx: &Context,
     rid: &Rid,
-    mut work: crate::message::MmWorkItem,
+    mut work: crate::message::request::MmWorkItem,
 ) -> Result<Vec<i32>, String> {
     let caller_hashes = std::mem::take(&mut work.mm_hashes);
     let input = super::payload::to_mm_input(work)?;
